+++ packages/framework/bundle/sandbox-base/bundle.js
@@ -56,13 +56,7 @@ function() {
 	"protocol" : ["Oskari.bundle.Bundle"],
 	"source" : {
 
-		"scripts" : [{
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/sandbox/sandbox.js"
-		}, {
-			"type" : "text/javascript",
-			"src" : "../../../../sources/framework/sandbox/sandbox-state-methods.js"
-		}],
+		"scripts" : [],
 		"resources" : []
 	},
 	"bundle" : {